tutorial/custom_local_server_MCP/requirements.txt
langchain_mcp_adapters
langchain-openai
langchain 
langchain-core 
langgraph
langgraph-cli[inmem]
langgraph-sdk
langchain_core 
langchain-anthropic
mem0ai 
protobuf==6.31.1 
grpcio 
grpcio-tools
langgraph-checkpoint-postgres
psycopg_binary